Fix callback crashes
The `CallVoidMethod` takes a variable argument list,
corresponding to the Java types the callee expects.

Unfortunately, since day one, we've been passing in the wrong type.
This just happened to work due to things randomly lining up that way,
but was a time bomb waiting to go off. With the switch to clang, it did.